silence clang analyzer warning
Clang points out that in the case of
0-dimensional meshes, this assignment
would be garbage.
It will probably be silenced if I
zero the initial set contents.

Although I could add an assert to inform
clang that 0D meshes can't happen,
the truth is weird stuff happens at SCOREC
and I shouldn't preclude the need to
temporarily write a 0D mesh to file.