adding extra check pointed by clang
clang now points out that I will compute
garbage if checking an empty set.
since check_set is only used in a case
where the set must not be empty, just
assert that the set is not empty.
This is a verification function, so in this
case clang is right, that check is needed.