got the code to pass clang static analysis
Some of the changes here are reiterations
of ones that were thought to have been
fixed by the previous few commits.
Overall, no actual bugs were found, it
was mainly a matter of clarifying our
intentions with asserts and checks.

The hack in parma_partinfo is unfortunate,
and only exists because people will pay
more attention to static analysis warnings
if the natural state is completely clean.