get copy alignment to work with local copies
The first version of this algorithm was designed
for off-part copies only, and was copied over
for matching. This works ok for simple matching,
but for multiple local copies it breaks down
because the sender doesn't know which downward
copies to send.

We alter the algorithm to send local downward
identifiers, and have the receiver compare
them to copies of its downward entities.
This is robust enough to handle multiply
periodic cases.