airstrike on compiler warnings
I was sent a log file of a compilation
on FreeBSD with all the nasty cousins
of -Wall enabled.
This is my attempt to fix the majority
of the resulting warnings.

Notable warnings which remain
include the use of zero-length
arrays at the end of structs (struct gmi_set)
and the use of the "%zu" conversion specifier
in printf-like functions to print size_t.