still trying to silence size_t print warnings
so all was well until we got to a 32-bit freebsd
box which now complains that unsigned long is
too big a specifier for a value that is actually
just unsigned.
now we just cast whatever the heck it is to long
and use the specifier for long.
didn't use unsigned long since its longer to
type in the casts and signed still holds all sane values.