fix SMB match IO inconsistency
There was a mistake made in early mds_smb code
that wrote matches using MDS_TYPES instead of SMB_TYPES.
We kept that around but now that we're adding an MDS type
it will blow up unless we fix it once and for all.
SMB version 3 is the tainted version, we will retain
read support for that version's ordering since a ton
of files have been made this way already,
but the new version 4 will properly organize things
by SMB_TYPES.
Also made a single constant for the current SMB version.