use the smallest moment of inertia in RIB
This fixes a bug that was exposed by
the fix introduced in:

4115c1972228715a15afcfa24d0ac35d688516bc

Before that, we were using the largest
moment of inertia as the normal to the
cutting plane and we had a broken
implementation of apf::eigen.

annoyingly, that code was actually giving
decent normal vectors.
however, once the eigenvalues/vectors were fixed,
it made it clear that the largest moment of
inertia is the opposite of what you want:
a rod's length is along its smallest
moment of inertia.

Along the way, remove the lonely little
rib_math files.

There is still a nagging question of whether
we want weights to factor into the
choice of normal, but this will be discussed
later.

This commit should fix parma/rib for most people.