implement patch expansion on QR failure
now, as part of the check for whether
we are done building a patch, the code
will collect the spatial locations of
patch points, form the Vandermonde
matrix and try to get its QR decomposition

if this succeeds, the patch builder terminates
and the SPR code picks up afterwards by
using the QR decomposition to fit polynomials
for each component of the field of interest.
this process is fast and guaranteed to succeed.

this requires storing the sample points
and QR decomposition temporarily in the
patch object.