adding non-threaded versions of split and zsplit
these are the main users of threading at
the moment, and multiple projects are
reporting annoyance at the thread requirement.
with PCU_Switch_Comm and MPI_Comm_split we
get the same effect.

the executable are transparently built
using one source file or another depending
on CMake's ENABLE_THREADS.
the test suite passes with both versions,
however, two caveats:

1) because the nothread versions can't
spawn threads, they need to be run
with the output number of parts, not the
input. testing.cmake handles this with
a macro.

2) tag/field transfer is not implemented
in the nothread versions