adding pcu_order system
this subsystem will capture
messages received by pcu_msg
and order them by sender,
then pretend to be the pcu_msg
and provide them to the user in
sorted order.

the purpose of this system is
to make PCU deterministic, in
turn helping a user application
to be deterministic in parallel.