don't let the boundary or core of a component be empty
A component with a single element had all of its vertices shared with a
component with a smaller id and thus had no vertices assigned to the component.
With this fix the component will still not own those vertices, but it will
have its boundary and core defined; which is required for distance computation.