changing much of the Chef element graph
1) ILWORKF is now ordered symmetrically
just like ILWORK, and we are *dropping*
the "element on the other side" numbers.
Consequently, it was much easier to
generalize getVertexLinks to build
links of mesh faces.
ph::Links is back to what it used to
be, and encodeILWORKF is responsible
for turning faces to elements and
getting their local IDs

2) IENNEIGH is formatted as before, but is now
the responsibility of phLinks.cc
and not apfInterElement.cc
previously, the parallel features
of apfInterElement.cc were used
to help construct ILWORK, but since
ph::getLinks has that job now,
the remaining code doesn't need
anything from apfInterElement.cc,
and if we make it phasta-specific
then we can generate the array
in FORTRAN order to begin with.