axisymmetric angle BC and iBC bit (untested)
this comes with lots of reorganizing of
the newly added model geometry logic
for match filtering, to reuse the
axisymmetric geometry math.

as with match filtering, there is now
a variable called "axisymmetry" in
adapt.inp for turning this code on
and off.

this is a backport from core-sim,
so in fact it is missing the crucial call
to attachAllAngleBCs
because thats in the model loading
function specific to Simmetrix.

this commit also fixes another uninit
issue in generating BC arrays
(and therefore will change the
CTest gold files)