always place vertices at edge midpoints
dynamic placement based on the size
field for many repeated refinements
results in ugly jumps in mesh resolution
along the boundaries of input elements.
always placing mid-edge vertices at
the parametric midpoint solves this.

there was also a ton of machinery to support
the various consequences of non-midpoint
vertices, which was also removed