reworking the basic containers package
note that although we may still replace
the code with the things showing up
in mth/, this commit just changes how
the containers fit into the build
system, and renames the package to "can".
they are now a full SCOREC package
and a Trilinos subpackage.
this is basically a template for
the way we will probably add components
from now on.