(squash) massive changes
In chronological order:

1) added an eigenvalue / eigenvector
decomposition function based on the
QR algorithm.
This includes Hessenberg reduction
and Wilkinson shifting, i.e. all the
bells and whistles.
Right now used for 3x3, but should work
for larger sizes.

2) Added a math compatibility file
apf/apf2mth.h, to contain
convenience functions for converting
between the math types.
This makes APF officially depend on
MTH, and so SPR doesn't need to list
an explicit dependency to MTH anymore.

3) replaced most of the APF math
in ParMA RIB splitting with MTH equivalents.
This is what the compatibility file was for.
Also, this includes switching out the
cubic roots and null space hand-waving
eigensolver for the new QR solver.
This is probably the most significant change.
Expect to see new things from RIB, hopefully
good things.

That is all.