apf: add ipnode to integrator
it is useful to know which integration point node the integrator is operating
on when atPoint() is called so that functions like getMatrix(field,
meshEntity, node, matrix) can be easily called. One should probably call the
corresponding function getMatrix(element, param, matrix) when possible,
but it does not appear possible to create apf::Element*'s for fields
like IPFields.