add different code for size field scaling
This code operates given only an isotropic
size field, making no assumptions on where
it came from, and scales it so that MeshAdapt
will produce roughly the target number of elements.
It was added at @onkarsahni's requested based
on the publication listed in the comments.

It is completely untested at this time.