two major changes in ma/maSnap.cc
1) "interpolateParametricCoordinates" now checks to make sure the
resulting point corresponding to the interpolated parameters is
actually inside the model (sometimes for periodic model faces
the resulting point ends up outside of the model). If not it
repeats the calculations with a slightly different logic.

2) "visualizeGeometricInfo" now adds an additional vector field
to the mesh that holds the difference (snap_to_x - before_snap_x).
Here snap_to_x is the location on the model boundary that the vertex
needs to snap to, and before_snap_x is the location before snapping.