add apf::clear()
This function is intended to remove
all content from an apf::Mesh, meaning
entities, fields, tags, etc.
It needs special support from MDS because
removing the entities one at a time leaves
the arrays in place and full of holes,
and the users of this new function need
new entities to appear in the order created.
This is implemented via apf::Mesh2::clear_()

The reason for this function is to have the
equivalent of apf::createEmptyMdsMesh(), without
changing the apf::Mesh pointer location.