phasta: write the mesh after writing detaching fields
Writing the mesh will reorder the entities if needed (i.e. the mesh was
modified and the arrays have 'holes').  Since we write the fields before
writing the mesh, the user must call 'balanceAndReorder' prior to
'preprocess' to ensure that the entity order in the phasta fields is
consistent with the mesh.