A better way to visualize higher-order meshes
writeCurvedWireFrame creates a separate wire-frame mesh (e.g., edges
only) and writes it in Paraview format.
Each edge in the wire-frame is subdivided to give the impression
of it being curved. This wire-frame mesh can be used (loaded into
paraview) on top of the mesh generated by the surface/volume
subdivisions (implemented previously).