Taking care of degenerate model faces
This commit takes care of degenerate model faces in the model's
parametric space, by correctly accounting for singularities (e.g. poles
in the sphere), see issue #52.
TODO: generalize this further to account for cases where one of the
poles (or both of them) are missing (e.g., a sphere in which one of the
poles is cut out)