support find_package with components
For example, if gmi_sim is not required then call

  find_package(SCOREC 2.1.0 REQUIRED OPTIONAL_COMPONENTS gmi_sim
               CONFIG PATHS ${SCOREC_PREFIX} NO_DEFAULT_PATH)

and the variable SCOREC_gmi_sim_FOUND will be set to true if found, and false
otherwise. Likewise, to require gmi_sim use the call

  find_package(SCOREC 2.1.0 REQUIRED gmi_sim
               CONFIG PATHS ${SCOREC_PREFIX} NO_DEFAULT_PATH)

and if gmi_sim is not found then a CMake error is emitted,
processing continues, but generation ( e.g.) Makefiles ) is skipped.