attempt to catch tag problem before segfault
If a tag or a field is created in a parma group code function
then the processes outside the group will not have the tag.  When the
mesh is later expanded to all processes a segfault will occur when a
process tries to access a tag that does not exist.  This assertion
detects when the sending process has more tags than the receiving process
and prints a helpful (?) message about tag creation.