Angle computation routines where added to apf.
This is supposed to work for both linear and curved simplexes. The
following angles are computable:
Edge-Edge
Edge-Face
Face-Face (if they share an edge).
This computations are based on columns of the Jacobian Matrix and that
is why they work for both linear and curved elements. Note that for
curved elements the Jacobian has to be computed as some "point". That
point is chosen according to the angle being requested. More
specifically
Edge-Edge (the point will be the shared vertex)
Edge-Face (the point will be the shared vertex)
Face-Face (the point will be the mid-point of the shared edge)