Updates to improve surface (2D) mesh adapt
This commit brings in 2 important changes
1- When there are problems with geometric similarity during snap,
instead of finding the parametric coordinates of the vertex at the mid-edge
point by using only the two edge verts, we use all the verts connected to
the vert-to-be-snaped by edges for interpolation.
2- Before this commit large angle triangles were fix by trying an edges-swap
on all the edges of the triangle in no specific order. Now we only try to swap
the edge opposite to the vertex with the largest angle (in metric space).