lots of changes to createEntity_
The most important being for the case of making a tet from its downward
faces. Since the Capstone api does not work for it we have to figure out
the downward verts to the tet using the downward faces. Since the
orientation of the tet is hard to determine we make to tets and pick the
one with positive volume. This would not have been needed had the api in
capstone worked for making a region from downward faces.