[ma] Add single-split-collapse functionality
[ma][sliver-fix] Replace sliver type table with 2-D array

Takes into account the fact that a vertex may be too close to an edge,
and that we might want to use a different fixer for it.

Two more types are added to that effect in `LargeAngleTetFixer`:
`EDGE_VERT` and `VERT_VERT`. Rotation in `matchSliver` is such that edge
v_0--v_1 is too close to v_3 for `EDGE_VERT` and v_0 and v_3 are too
close for `VERT_VERT`.

[ma][sliver-fix] Add proximity to edges in `getSliverCode`

Currently, edge-vert and vert-vert cases are treated the same as face-vert.

[ma][sliver-fix] Logic for flat first face

Similar code structure, except using projection of third vertex of first face onto the opposite edge.

[ma][sliver-fix] Using face quality condition to choose identifiers

Also tweaked the threshold for closeness,

[ma][sliver-fix] Insert single split collapse into build

[ma][sliver-fix] Insert single-split-collapse in large-angle-fixer

Also removed some outputs used for debugging.