A new test adapting on an analytic sphere model
In this test a model of a sphere is made, first. Then a coarse mesh is
made that has only two tets and all the entities are classified on the
model. The adapt called with requesting for 2 levels of uniform
refinement to demonstrate that snapping work fine given an analytic
model.