Updated maSnapper.cc
The static routine updateVertexParametricCoords now computes the
parametric coordinates of the vertex based on parametric coordinates of
the mid points of the upward adjacent edges that are on the same model
face. Previously, this was done by using the parametric coordinates of
the vertexes connected to this vertex, and that was causing problems
when the underlying model face was periodic.