Initial commit for python-wrappers
This commits does the wrappers for some of the PCU APIs

For now everything is done using swig. That is "swig" is call
automatically during cmake configure process and _.so (python Module)
files are build during the make process.

In the future it might be beneficial to also include build instructions
using the "setup.py" approach.