Takes care of blending when doing Bezier Fields
Note that blending is used internally in Bezier Shapes. More
specifically when a BezierCurver object is set with a non-zero third
argument, Bezier shapes for tets and triangles might behave differently.
Therefore, we needed to make sure the blending orders are set to zero at
the beginning of the convertInterpolationgFieldsToBezier routine and
then reverted back to the original values at the end.

There still seems to be necessary to construct a new BezierCurver object
with last argument 0 for the bezier fields to work properly.