Uses templates for the order P
The drawback of this approach is that different order nedelec shapes has
to be specialized manually. But since this is done inside getNedlec
routine, we can just specialize as many as we want and clients don't
have to worry about this. For now the specializations are done for
orders 1 to 10.