Updates the reorderData
The previous version would not negate the value of the 0 index. So we
have to follow a notation similar to mfem. For example in the case of
edges in a tet we have

different local/global dirs     same local/global dirs
........., -3, -2, -1        |         0, 1, 2, ..........
where
the value is the same at 0 and -1 except for a minus sign
the value is the same at 1 and -2 except for a minus sign
the value is the same at 2 and -3 except for a minus sign
and so on.

alignSharedNodes in apfNedelec EntityShapes have to be updated to
account for this.