Adds another getElementNodeXis
This one returns the xi in the order the appear in the downward adjacent
entities. This a more useful one for the purposes of field transfer from
solvers and should eliminated the need for calls to aligneShareNodes
(which is a very low level functionality in apf Shapes and its better
that the clients of the code would not use them).