Fixes rotateTetXi
Previously this was implemented exactly the same way as unrotateTetXi,
which was not right. The current implementation does the intended task,
which is basically the inverse of unrotateTetXi. Therefore, given a xi
vector, calling unrotateTetXi(xi, rotation) and then rotateTetXi(xi,
rotation) should make xi to have its original coordinates.