Fixes a bug in getElementData
Previously, for efficiency reasons, the size of the intermediate array
"adata" was set ot "nen" (the total number of nodes associated with the
entity). This is not enough for cases where the number of components for
the fields is large. For example for a cubic triangle nen is 10 (3 vertex nodes,
6 edge nodes, and 1 face node). Now for the same triangle "adata" for each edge
will need the min size of 18 if the underlying field type is matrix (9 components
per node x 2 nodes).