Add state to PCU mpi interface.
This commit adds a MPI state that's separate from the global variables
to the internal pcu_mpi interfaces. The user level PCU interface remains
unchanged. However, this change lays the groundwork for PCU2 that will
not make use of global state. This is important for some use cases like
working around the spectrum-mpi comm_dup bug. The removal of global
state also improves testability and modularity of this code.