simplify and improve logic for setting C++ standard
Added a 'bob' cmake helper function to set the C++ standard. The
modifications do not change the default behavior of using C++11. The
user is given the ability to increase the standard to > C++11, or enable
the use of extensions, by passing CMAKE_CXX_STANDARD and
CMAKE_CXX_EXTENSIONS, respectively, to cmake.

Following approach given in "Professional CMake: 19th Edition".